+++ src/components/Main/PlaceComponent.tsx
@@ -5,14 +5,14 @@ export default function PlaceComponent({ places }: PlaceComponentProps) {
   const navigate = useNavigate();
 
   return (
-    <div className="grow flex flex-col px-20 justify-center h-[480px] bg-dftBackgroundGray">
+    <div className="grow flex flex-col px-80 justify-center bg-dftBackgroundGray">
       <div>
-        <h2 className="text-2xl font-custom text-center tracking-tight text-gray-900 sm:text-4xl">
+        <h2 className="text-2xl font-main text-center tracking-tight text-gray-900 sm:text-4xl">
           여기저기 여행을 시작해보세요!
         </h2>
-        <p className="mt-4 font-custom text-center text-gray-500">
+        <p className="mt-4 mb-4 font-main text-center text-gray-500">
           지금 대한민국에서{' '}
-          <span className="text-red-300">인기 많은 여행지</span>는 어디일까요?
+          <span className="text-red-300 ">인기 많은 여행지</span>는 어디일까요?
         </p>
       </div>
       <div className="grid grid-cols-4 gap-4 sm:gap-6 lg:gap-8">
