+++ src/components/all/Footer.tsx
@@ -1,6 +1,6 @@
 export default function Footer() {
   return (
-    <footer className="flex justify-center items-center h-[80px]">
+    <footer className="flex justify-center bg-dftBackgroundGray h-[80px] mt-32">
       <img
         src="/assets/EnglishLogo.svg"
         alt="This is english here and there logo!"
