+++ tailwind.config.js
@@ -4,12 +4,16 @@ export default {
   theme: {
     extend: {
       fontFamily: {
-        custom: ['Gugi', 'sans-serif'],
+        main: ['Gugi', 'sans-serif'],
+        text: ["Nanum Gothic", 'sans-serif'],
       },
       colors: {
         dftBackgroundGray: '#F4F5FF',
         logoOrange: '#FAAC64',
         logoBlue: '#93BCFA',
+        textRed: '#FF8577',
+        textGray: '#BBB8B8',
+        chatRoomPurple: 'rgba(85, 81, 255, 0.90)',
       },
     },
   },
